feat: Implémentation i18n, mise à jour dépendances et refactoring
