+++ messages/en.json
@@ -1,0 +1,9 @@
+{
+  "HomePage": {
+    "welcome": "Welcome to BrainBloom",
+    "getStarted": "Get Started",
+    "prompt": "Ready to plant your first idea?",
+    "description": "Your digital garden of ideas. Cultivate thoughts, connect concepts, and watch your knowledge bloom."
+  }
+}
+
