+++ messages/fr.json
@@ -1,0 +1,9 @@
+{
+  "HomePage": {
+    "welcome": "Bienvenue sur BrainBloom",
+    "getStarted": "Commencer",
+    "prompt": "Prêt(e) à planter votre première idée ?",
+    "description": "Votre jardin numérique d'idées. Cultivez vos pensées, connectez les concepts et regardez vos connaissances s'épanouir."
+  }
+}
+
